charging mode: Fix boot loop on USB
-This problem was simply a matter of poor timing. When the device
boots while connected to a USB cable; plugged into, for example, a
PC, the bootloader kicks the device on. The adc_battery driver comes
online and its cable "connect/disconnect" callback is fired as
soon as it registers. Since we initialize with a value of CABLE_TYPE_NONE
and the callback is fired before the USB Switch driver "fsa9480" has been
intialized and because are in a "charging_mode_booting" situation; the adc_battery
"thinks" we have unplugged the cable while in charging mode.
BOOTLOOP...

-This solution here is to initialize with CABLE_TYPE_USB and let the
USB Switch driver put the proper value in when it is initialized.